backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
requests>=2.31.0
python-dotenv>=1.0.0